sometimes does some off stuff with rendeing but debounce is working nicely
