+++ packages/web/src/brandkit/KeyImagery.tsx
@@ -92,7 +92,7 @@ const Illustrations = React.memo(
                     key={illo.name}
                     description={illo.description}
                     name={illo.name}
-                    preview={{ uri: illo.preview }}
+                    preview={illo.preview}
                     uri={illo.uri}
                     loading={false}
                     size={size}
